requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
qrcode[pil]==8.0
ReportLab==4.2.5
Pillow==11.1.0
Werkzeug==3.1.3
gunicorn==23.0.0
